Init for Funds by Region
I copy-pasted the Top100Page stuff, which currently needs error handling
for its api calls. These changes will have to be propogated to the Funds
by Region page once they are done.

We currently are only going to store data about North America, so it's
hardcoded in. we will have to find some way to sort the funds before
they are displayed.